cleanup: move deployment logic out of morph.go
A lot is moved into the new package 'cruft'. This is code I believe will
have to be refactored heavily later on, thus the name of the package.
