+++ cruft/secrets.go
@@ -1,0 +1,134 @@
+package cruft
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/DBCDK/morph/common"
+	"github.com/DBCDK/morph/healthchecks"
+	"github.com/DBCDK/morph/nix"
+	"github.com/DBCDK/morph/secrets"
+	"github.com/DBCDK/morph/ssh"
+	"github.com/DBCDK/morph/utils"
+)
+
+func ExecListSecrets(hosts []nix.Host) {
+	for _, host := range hosts {
+		singleHostInList := []nix.Host{host}
+		for _, host := range singleHostInList {
+			fmt.Fprintf(os.Stdout, "Secrets for host %s:\n", host.Name)
+			for name, secret := range host.Secrets {
+				fmt.Fprintf(os.Stdout, "%s:\n- %v\n", name, &secret)
+			}
+			fmt.Fprintf(os.Stdout, "\n")
+		}
+	}
+}
+
+func ExecListSecretsAsJson(opts *common.MorphOptions, hosts []nix.Host) error {
+	deploymentDir, err := filepath.Abs(filepath.Dir(opts.Deployment))
+	if err != nil {
+		return err
+	}
+	secretsByHost := make(map[string](map[string]secrets.Secret))
+
+	for _, host := range hosts {
+		singleHostInList := []nix.Host{host}
+		for _, host := range singleHostInList {
+			canonicalSecrets := make(map[string]secrets.Secret)
+			for name, secret := range host.Secrets {
+				sourcePath := utils.GetAbsPathRelativeTo(secret.Source, deploymentDir)
+				secret.Source = sourcePath
+				canonicalSecrets[name] = secret
+			}
+			secretsByHost[host.Name] = canonicalSecrets
+		}
+	}
+
+	jsonSecrets, err := json.MarshalIndent(secretsByHost, "", "  ")
+	if err != nil {
+		return err
+	}
+	fmt.Fprintf(os.Stdout, "%s\n", jsonSecrets)
+
+	return nil
+}
+
+func ExecUploadSecrets(opts *common.MorphOptions, hosts []nix.Host, phase *string) error {
+	sshContext := ssh.CreateSSHContext(opts)
+
+	for _, host := range hosts {
+		if host.BuildOnly {
+			fmt.Fprintf(os.Stderr, "Secret upload is disabled for build-only host: %s\n", host.Name)
+			continue
+		}
+		singleHostInList := []nix.Host{host}
+
+		err := secretsUpload(opts, sshContext, singleHostInList, phase)
+		if err != nil {
+			return err
+		}
+
+		if !opts.SkipHealthChecks {
+			err = healthchecks.PerformHealthChecks(sshContext, &host, opts.Timeout)
+			if err != nil {
+				fmt.Fprintln(os.Stderr)
+				fmt.Fprintln(os.Stderr, "Not uploading to additional hosts, since a host health check failed.")
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+func secretsUpload(opts *common.MorphOptions, ctx ssh.Context, filteredHosts []nix.Host, phase *string) error {
+	// upload secrets
+	// relative paths are resolved relative to the deployment file (!)
+	deploymentDir := filepath.Dir(opts.Deployment)
+	for _, host := range filteredHosts {
+		fmt.Fprintf(os.Stderr, "Uploading secrets to %s (%s):\n", host.Name, host.TargetHost)
+		postUploadActions := make(map[string][]string, 0)
+		for secretName, secret := range host.Secrets {
+			// if phase is nil, upload the secrets no matter what phase it wants
+			// if phase is non-nil, upload the secrets that match the specified phase
+			if phase != nil && secret.UploadAt != *phase {
+				continue
+			}
+
+			secretSize, err := secrets.GetSecretSize(secret, deploymentDir)
+			if err != nil {
+				return err
+			}
+
+			secretErr := secrets.UploadSecret(ctx, &host, secret, deploymentDir)
+			fmt.Fprintf(os.Stderr, "\t* %s (%d bytes).. ", secretName, secretSize)
+			if secretErr != nil {
+				if secretErr.Fatal {
+					fmt.Fprintln(os.Stderr, "Failed")
+					return secretErr
+				} else {
+					fmt.Fprintln(os.Stderr, "Partial")
+					fmt.Fprint(os.Stderr, secretErr.Error())
+				}
+			} else {
+				fmt.Fprintln(os.Stderr, "OK")
+			}
+			if len(secret.Action) > 0 {
+				// ensure each action is only run once
+				postUploadActions[strings.Join(secret.Action, " ")] = secret.Action
+			}
+		}
+		// Execute post-upload secret actions one-by-one after all secrets have been uploaded
+		for _, action := range postUploadActions {
+			fmt.Fprintf(os.Stderr, "\t- executing post-upload command: "+strings.Join(action, " ")+"\n")
+			// Errors from secret actions will be printed on screen, but we won't stop the flow if they fail
+			ctx.CmdInteractive(&host, opts.Timeout, action...)
+		}
+	}
+
+	return nil
+}
